+++ lua/blak/config/defaults.lua
@@ -40,6 +40,10 @@ return {
     provider = "fff", -- fff | snacks | telescope | fzf_lua
   },
 
+  explorer = {
+    provider = "oil", -- oil | snacks
+  },
+
   snacks = {},
 
   treesitter = {