+++ lua/blak/extras/init.lua
@@ -14,6 +14,7 @@ local modules = {
   "blak.extras.git.diffview",
   "blak.extras.ai.copilot",
   "blak.extras.editor.neotree",
+  "blak.extras.editor.snacks_explorer",
   "blak.extras.editor.telescope",
   "blak.extras.editor.fzf_lua",
 }